requirements.txt
requests==2.31.0
dataclasses-json==0.6.4
python-dateutil==2.9.0.post0